Don't expand formatted ranges to include neighbouring tokens
If the formatted range starts or ends on whitespace, previously we would expand
the range to the next token. This meant that if the formatted range
started/ended on whitespace or comments, neighbouring statements owuld also get
reformatted.

This change causes the formatter to contract the range instead.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=95241146