Fix partial formatting bug regions endling on newlines.
Formatting a range that includes a terminating newline shouldn't cause the next
statement to be formatted.

This changes the formatter to treat newlines as trailing trivia of the last
token on the previous line, instead of leading trivia of the first token on the
next line.
-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=105326621