Refactor FormatFileCallable.call
prior to adding logic to it to do import ordering if requested on the
command-line. Also reformat this file using gjf itself.

This is just moving stuff around to make the logic of the method easier
to follow. It does fix a small bug where the wrong message would be
output if writing stdout produced an IOException. That can't actually
happen now because the Appendable in question is actually a PrintWriter.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=117545448