Support running formatter thru a Maven plugin
Not all the CLI options are available. Configuration parameters
include 'style' and 'failOnError'. Refer to README.md for usage.

This is an alpha version. It brings a few TODO points as suggestions
to be added to core API. Configurations can be extended to allow
for example `excludes` and `includes`.

The plugin defines a profile 'run-self' to apply formatting to the
plugin source itself. To run first `mvn install` then `mvn verify -Prun-self`.

For context and discussion, please refer to #69