Fix newline normalization
The handling of \n tokens was inconsistent for newlines in comments blocks
at the end of files, which was causing the wrong column to be reported,
which caused incorrect line comment wrapping.

The logic here mimics the equivalent newline handling in the
`if (tokenOp.realOrImaginary().isReal()) {` branch up above.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=165230368