Allow partial formatting within switch statement groups
Previously we tried to be clever about reformatting entire statement
groups, but this works out badly for large statement groups. This
increases the risk of surprising partial formatting results when
formatting a single line in a mis-indented switch, since as usual the
formatted line will be indented relative to the fully formatted file and
not the surrounding lines.

MOE_MIGRATED_REVID=171983598