Put empty class and member declarations on their own line
Per JLS 7.6 and 8.2 `;` is a valid empty top-level type or class member
declaration. This change puts it on its own line, instead of adding it
at the end of the preceding line.

Omitting the `;` is recommended: the goal of this change is just avoid a
partial formatting bug, and to call attention to the unnecessary element.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=235586393