Update Eclipse plugin build
Updates the Eclipse plugin build logic. The logic is now completely
separate from the main google java format build. Instead of relying on
the build artifacts of the core build, the needed google-java-format and
guava build artifacts used for the build are now pulled from maven.

Updates google-java-format version used for the build to '1.8', the
build JDK to 11, and removes old dependencies which are no longer used.

Updates tycho version to 1.7.0 in order to build with JDK > 8.

Updates the build guide in the README.

Co-authored-by: Kelvin Glaß <kelvin.glass@fu-berlin.de>
Co-authored-by: Tobias Bouschen <tobias.bouschen@googlemail.com>