Break out Java 15-specific code to java15 package
The previous commit didn’t consider backwards compatibility with regards to
compilation on a pre-Java 15 JDK.  This patch breaks out all Java 15-specific
code to the com.google.googlejavaformat.java.java15 package.

This change requires us to make ModifierOrderer non-final and some of its
methods non-static so that we can override asModifier to handle sealed and
non-sealed in the new Java15ModifierOrderer.

The new Java15InputAstVisitor overrides a new protected method,
getPermitsClause, that reflectively invokes getPermitsClause.  Extracting this
method seems to make the most sense, as visitClassDeclaration is quite complex
and overriding it as a whole didn’t seem like a good idea.

Finally, creation of the InputAstVisitor was simplified, as creating it via
reflection for Java14 and Java15 isn’t necessary, as Java14InputAstVisitor and
Java15InputAstVisitor are built on earlier JDKs anyway.