Explicitly state GFJ version to use for Eclipse plugin build
Re-adds the GJF version specification to the Eclipse plugin pom.

This decouples the GJF version from the Eclipse plugin version. As a
result, it is possible to increase the Eclipse plugin version without
changing the used GJF core version (e.g. in cases where you want to
increase the patch level of the plugin version when deploying fixes for
the Eclipse plugin).

Additionally, it allows for easier builds with local snapshot versions
of GJF.

Also updates the README to link to the current release. Removes explicit
mention of the Eclipse plugin release 1.6. Links to the release page
instead of a specific JAR file so that the link does not have to be
updated with every release.