Format type annotation as part of the type, not part of the modifiers list
Given e.g. `@Deprecated @Nullable Object foo() {}`, prefer this:

```
@Deprecated
@Nullable Object foo() {}
```

instead of:

```
@Deprecated
@Nullable
Object foo() {}
```

The implementation is complicated by the fact that the AST doesn't store
source position information for modifiers, and there's no requirement that
declaration annotations, modifiers, and type annotations appear in any
particular order in source.

To work around this, we examine the token stream to figure out the ordering
of the modifiers and annotations.

https://github.com/google/google-java-format/issues/5

PiperOrigin-RevId: 392769609