Automated rollback of commit 1a875792f6a074f3f0a504ddb8c7cd932ba6b073.
*** Reason for rollback ***

TAP shows this is broke Voice Access (and perhaps other projects)

[]

*** Original change description ***

Format type annotation as part of the type, not part of the modifiers list

Given e.g. `@Deprecated @Nullable Object foo() {}`, prefer this:

```
@Deprecated
@Nullable Object foo() {}
```

instead of:

```
@Deprecated
@Nullable
Object foo() {}
```

The implementation is complicated by the fact that the AST doesn't store
source position information for modifiers, and there's no requirement that
declaration annotations, modifiers, and type annotations appear in any
particular order in source.

To...

***

PiperOrigin-RevId: 392785771