Don't use single-line for for javadoc without a summary fragment
Using the multi-line form here may call attention to the missing
summary fragment.

The style guide can also be read as discouraging the single-line
form for non-required javadoc missing a summary fragment
(see b/74239048#comment4), but I'm not sure that's the intent,
and I'm looking at changing it (unknown commit).

PiperOrigin-RevId: 418174535