Update the `--add-exports` instructions.
I don't really understand _why_ this works. I'm able to (easily) reproduce the
problem by not including the `=` in my `idea64.options` file. But then if I
copy/paste the command line out of `ps`, it works fine.

And since it's easy enough to change this, I'm inclined to just do that?

Fixes #908.

PiperOrigin-RevId: 514440074