Update text block formatting after nickreid's improvements in unknown commit
Previously the formatting was adding a forced break at the beginning of text blocks, which caused issues like #1081. With the changes in the baseline CL it correctly handles the 'width' of text blocks containing newlines as infinity, instead of counting the number of characters and treating the newline as having width 1.

Fixes https://github.com/google/google-java-format/issues/1081

PiperOrigin-RevId: 620933964