Always run integration test for #1161
The intent was to only run the test on JDK 23+ and to disable it until the linked issue was fixed, but it actually resulted in the test being discovered and run on all JDK versions.

That's fine, because the test is backwards compatible (`///` comments are valid in all versions, they just aren't markdown comments until 23).

The test is being fixed in https://github.com/google/google-java-format/pull/1161

PiperOrigin-RevId: 674352956