Fix logic for checking if text blocks are already deindented
Check the last line instead of the first line of the text block contents. Empty lines in the text block have no leading whitespace but do not indicate the contents are deindented, only non-empty lines should be considered. The last line will never be empty, it contains at least the close delimiter.

PiperOrigin-RevId: 697785256